pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
prisma = "^0.15.0"
ipykernel = "^6.29.5"
dspy-ai = "^2.4.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "scripts.dev:run"